add common Tooltip and use it for FlamegraphTooltip
