+++ packages/pyroscope-flamegraph/src/Tooltip/LeftClickIcon.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+export default function LeftClickIcon() {
+  return (
+    <svg id="Layer_1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30">
+      <defs />
+      <path
+        style={{ fill: '#7d7d7f' }}
+        d="M24,19.56c0,5.22-3.54,9.44-7.92,9.44h-3.17c-4.37,0-7.92-4.23-7.92-9.44v-7.56H24v7.56Z"
+      />
+      <path
+        style={{ fill: '#7d7d7f' }}
+        d="M15,1h.82c4.52,0,8.18,3.73,8.18,8.33v1.67H15V1Z"
+      />
+      <path d="M13.18,1h.82V11H5v-1.67c0-4.6,3.66-8.33,8.18-8.33Z" />
+    </svg>
+  );
+}
