+++ packages/pyroscope-flamegraph/src/Tooltip/Tooltip.module.scss
@@ -1,4 +1,4 @@
-.flamegraphTooltip {
+.tooltip {
   width: 420px;
   position: fixed;
   background: var(--ps-tooltip-bg);
@@ -13,7 +13,7 @@
     width: 450px;
   }
 
-  .flamegraphTooltipName {
+  .tooltipName {
     color: var(--ps-tooltip-text);
     background-color: var(--ps-tooltip-header-bg);
     border-top-right-radius: 4px;
@@ -57,13 +57,14 @@
     }
   }
 
-  .rightClickInfo {
+  .clickInfo {
     margin-bottom: 10px;
     display: flex;
     justify-content: center;
     color: var(--ps-right-click-info);
 
     svg {
+      color: initial;
       display: block;
       align-self: center;
       width: 22px;
