+++ packages/pyroscope-flamegraph/src/Tooltip/Tooltip.tsx
@@ -1,56 +1,21 @@
-import React from 'react';
-import { Maybe } from 'true-myth';
+import React, {
+  CSSProperties,
+  RefObject,
+  ReactNode,
+  useEffect,
+  useState,
+  useRef,
+  useCallback,
+  Dispatch,
+  SetStateAction,
+} from 'react';
 import clsx from 'clsx';
 import type { Units } from '@pyroscope/models/src';
-import type { Unwrapped } from 'true-myth/maybe';
-import {
-  getFormatter,
-  numberWithCommas,
-  formatPercent,
-  ratioToPercent,
-} from '../format/format';
+
 import RightClickIcon from './RightClickIcon';
-import {
-  FlamegraphPalette,
-  DefaultPalette,
-} from '../FlameGraph/FlameGraphComponent/colorPalette';
+import LeftClickIcon from './LeftClickIcon';
+
 import styles from './Tooltip.module.scss';
-
-type xyToDataSingle = (
-  x: number,
-  y: number
-) => Maybe<{ format: 'single'; name: string; total: number }>;
-
-type xyToDataDouble = (
-  x: number,
-  y: number
-) => Maybe<{
-  format: 'double';
-  name: string;
-  totalLeft: number;
-  totalRight: number;
-  barTotal: number;
-}>;
-
-export type TooltipProps = {
-  canvasRef: React.RefObject<HTMLCanvasElement>;
-
-  units: Units;
-  sampleRate: number;
-  numTicks: number;
-  leftTicks: number;
-  rightTicks: number;
-
-  palette: FlamegraphPalette;
-} & (
-  | { format: 'single'; xyToData: xyToDataSingle }
-  | {
-      format: 'double';
-      leftTicks: number;
-      rightTicks: number;
-      xyToData: xyToDataDouble;
-    }
-);
 
 const tooltipTitles: Record<
   Units,
@@ -90,15 +55,42 @@ const tooltipTitles: Record<
   },
 };
 
-type TooltipData = {
-  units: Units;
-  percent: string | number;
-  samples: string;
-  formattedValue: string;
-};
+interface TooltipProps {
+  // canvas or table ref
+  dataSourceRef: RefObject<HTMLCanvasElement | any>;
 
-export default function Tooltip(props: TooltipProps) {
-  const { format, canvasRef, xyToData } = props;
+  // footer
+  shouldShowFooter?: boolean;
+  clickInfoSide?: 'left' | 'right';
+
+  setTooltipContent: (
+    setContent: Dispatch<
+      SetStateAction<{
+        title: {
+          text: string;
+          diff: {
+            text: string;
+            color: string;
+          };
+        };
+        tooltipData: TooltipData[];
+      }>
+    >,
+    onMouseOut: () => void,
+    e: MouseEvent
+  ) => void;
+  // for styles
+  type: 'flamegraph' | 'table';
+}
+
+export function Tooltip({
+  dataSourceRef,
+  shouldShowFooter = true,
+  clickInfoSide,
+  setTooltipContent,
+  type,
+}: TooltipProps) {
+  const tooltipRef = useRef<HTMLDivElement>(null);
   const [content, setContent] = React.useState({
     title: {
       text: '',
@@ -109,30 +101,23 @@ export default function Tooltip(props: TooltipProps) {
     },
     tooltipData: [] as TooltipData[],
   });
+  const [style, setStyle] = useState<CSSProperties>();
 
-  const [style, setStyle] = React.useState<React.CSSProperties>();
-  const tooltipEl = React.useRef<HTMLDivElement>(null);
-
-  const { numTicks, sampleRate, units, leftTicks, rightTicks, palette } = props;
   const onMouseOut = () => {
     setStyle({
       visibility: 'hidden',
     });
   };
 
-  // recreate the callback when the dependency changes
-  // that's to evict stale props
-  const memoizedOnMouseMove = React.useCallback(
+  const memoizedOnMouseMove = useCallback(
     (e: MouseEvent) => {
-      const formatter = getFormatter(numTicks, sampleRate, units);
-
-      if (!tooltipEl || !tooltipEl.current) {
+      if (!tooltipRef || !tooltipRef.current) {
         throw new Error('Missing tooltipElement');
       }
 
       const left = Math.min(
         e.clientX + 12,
-        window.innerWidth - tooltipEl.current.clientWidth - 20
+        window.innerWidth - tooltipRef.current.clientWidth - 20
       );
       const top = e.clientY + 20;
 
@@ -142,138 +127,68 @@ export default function Tooltip(props: TooltipProps) {
         visibility: 'visible',
       };
 
-      const opt = xyToData(e.offsetX, e.offsetY);
-      let data: Unwrapped<typeof opt>;
-
-      // waiting on
-      // https://github.com/true-myth/true-myth/issues/279
-      if (opt.isJust) {
-        data = opt.value;
-      } else {
-        onMouseOut();
-        return;
-      }
-
-      // set the content
-      switch (data.format) {
-        case 'single': {
-          const newLeftContent: TooltipData = {
-            percent: formatPercent(data.total / numTicks),
-            samples:
-              units === 'trace_samples' ? '' : numberWithCommas(data.total),
-            units,
-            formattedValue: formatter.format(data.total, sampleRate),
-          };
-
-          setContent({
-            title: {
-              text: data.name,
-              diff: {
-                text: '',
-                color: '',
-              },
-            },
-            tooltipData: [newLeftContent],
-          });
-
-          break;
-        }
-
-        case 'double': {
-          if (format === 'single') {
-            throw new Error(
-              "props format is 'single' but it has been mapped to 'double'"
-            );
-          }
-
-          const d = formatDouble(
-            {
-              formatter,
-              sampleRate,
-              totalLeft: data.totalLeft,
-              leftTicks,
-              totalRight: data.totalRight,
-              rightTicks,
-              title: data.name,
-              units,
-            },
-            palette
-          );
-
-          setContent({
-            title: d.title,
-            tooltipData: d.tooltipData,
-          });
-
-          break;
-        }
-        default:
-          throw new Error(`Unsupported format:'`);
-      }
-
+      setTooltipContent(setContent, onMouseOut, e);
       setStyle(style);
     },
 
     // these are the dependencies from props
     // that are going to be used in onMouseMove
-    [numTicks, sampleRate, units, format, xyToData]
+    [tooltipRef, setTooltipContent]
   );
 
-  React.useEffect(() => {
-    // use closure to "cache" the current canvas reference
+  useEffect(() => {
+    // use closure to "cache" the current dataSourceRef(canvas/table) reference
     // so that when cleaning up, it points to a valid canvas
     // (otherwise it would be null)
-    const canvasEl = canvasRef.current;
-    if (!canvasEl) {
+    const dataSourceEl = dataSourceRef.current;
+    if (!dataSourceEl) {
       return () => {};
     }
 
     // watch for mouse events on the bar
-    canvasEl.addEventListener('mousemove', memoizedOnMouseMove);
-    canvasEl.addEventListener('mouseout', onMouseOut);
+    dataSourceEl.addEventListener('mousemove', memoizedOnMouseMove);
+    dataSourceEl.addEventListener('mouseout', onMouseOut);
 
     return () => {
-      canvasEl.removeEventListener('mousemove', memoizedOnMouseMove);
-      canvasEl.removeEventListener('mouseout', onMouseOut);
+      dataSourceEl.removeEventListener('mousemove', memoizedOnMouseMove);
+      dataSourceEl.removeEventListener('mouseout', onMouseOut);
     };
-  }, [canvasRef.current, memoizedOnMouseMove]);
+  }, [dataSourceRef.current, memoizedOnMouseMove]);
 
   return (
     <div
-      role="tooltip"
-      data-testid="flamegraph-tooltip"
-      className={clsx(styles.flamegraphTooltip, {
+      data-testid="tooltip"
+      className={clsx(styles.tooltip, {
         [styles.flamegraphDiffTooltip]: content.tooltipData.length > 1,
       })}
       style={style}
-      ref={tooltipEl}
+      ref={tooltipRef}
     >
-      <div
-        className={styles.flamegraphTooltipName}
-        data-testid="flamegraph-tooltip-title"
-      >
+      <div className={styles.tooltipName} data-testid="tooltip-title">
         {content.title.text}
       </div>
-      <div
-        className={styles.functionName}
-        data-testid="flamegraph-tooltip-function-name"
-      >
+      <div className={styles.functionName} data-testid="tooltip-function-name">
         {content.title.text}
       </div>
-
       {content.title.diff.text.length > 0 ? (
         <TooltipTable data={content.tooltipData} diff={content.title.diff} />
       ) : (
         <TooltipTable data={content.tooltipData} />
       )}
-
-      <div className={styles.rightClickInfo}>
-        <RightClickIcon />
-        <span>Right click for more node viewing options</span>
-      </div>
+      <TooltipFooter
+        shouldShowFooter={shouldShowFooter}
+        clickInfoSide={clickInfoSide}
+      />
     </div>
   );
 }
+
+export type TooltipData = {
+  units: Units;
+  percent: string | number;
+  samples: string;
+  formattedValue: string;
+};
 
 function TooltipTable({
   data,
@@ -290,7 +205,7 @@ function TooltipTable({
 
   return (
     <table
-      data-testid="flamegraph-tooltip-table"
+      data-testid="tooltip-table"
       className={clsx(styles.tooltipTable, {
         [styles.tooltipDiffTable]: comparisonData,
       })}
@@ -315,7 +230,7 @@ function TooltipTable({
               <td>
                 {diff && (
                   <span
-                    data-testid="flamegraph-tooltip-diff"
+                    data-testid="tooltip-diff"
                     style={{ color: diff.color }}
                   >
                     {diff.text}
@@ -350,97 +265,37 @@ function TooltipTable({
   );
 }
 
-interface Formatter {
-  format(samples: number, sampleRate: number): string;
-}
+function TooltipFooter({
+  shouldShowFooter,
+  clickInfoSide,
+}: {
+  shouldShowFooter: boolean;
+  clickInfoSide?: 'left' | 'right';
+}) {
+  let clickInfo: ReactNode;
 
-function formatDouble(
-  {
-    formatter,
-    sampleRate,
-    totalLeft,
-    leftTicks,
-    totalRight,
-    rightTicks,
-    title,
-    units,
-  }: {
-    formatter: Formatter;
-    sampleRate: number;
-    totalLeft: number;
-    leftTicks: number;
-    totalRight: number;
-    rightTicks: number;
-    title: string;
-    units: Units;
-  },
-  palette: FlamegraphPalette = DefaultPalette
-): {
-  tooltipData: TooltipData[];
-  title: {
-    text: string;
-    diff: {
-      text: string;
-      color: string;
-    };
-  };
-} {
-  const leftRatio = totalLeft / leftTicks;
-  const rightRatio = totalRight / rightTicks;
-
-  const leftPercent = ratioToPercent(leftRatio);
-  const rightPercent = ratioToPercent(rightRatio);
-
-  const newLeft: TooltipData = {
-    percent: `${leftPercent}%`,
-    samples: numberWithCommas(totalLeft),
-    units,
-    formattedValue: formatter.format(totalLeft, sampleRate),
-  };
-
-  const newRight: TooltipData = {
-    percent: `${rightPercent}%`,
-    samples: numberWithCommas(totalRight),
-    units,
-    formattedValue: formatter.format(totalRight, sampleRate),
-  };
-
-  const totalDiff = percentDiff(leftPercent, rightPercent);
-
-  let tooltipDiffColor = '';
-  if (totalDiff > 0) {
-    tooltipDiffColor = palette.badColor.rgb().string();
-  } else if (totalDiff < 0) {
-    tooltipDiffColor = palette.goodColor.rgb().string();
+  switch (clickInfoSide) {
+    case 'right':
+      clickInfo = (
+        <>
+          <RightClickIcon />
+          <span>Right click for more node viewing options</span>
+        </>
+      );
+      break;
+    case 'left':
+      clickInfo = (
+        <>
+          <LeftClickIcon />
+          <span>Click to highlight node in flamegraph</span>
+        </>
+      );
+      break;
+    default:
+      clickInfo = '<TBD ?>';
   }
 
-  let tooltipDiffText = '';
-  if (!totalLeft) {
-    // this is a new function
-    tooltipDiffText = '(new)';
-  } else if (!totalRight) {
-    // this function has been removed
-    tooltipDiffText = '(removed)';
-  } else if (totalDiff > 0) {
-    tooltipDiffText = `(+${totalDiff.toFixed(2)}%)`;
-  } else if (totalDiff < 0) {
-    tooltipDiffText = `(${totalDiff.toFixed(2)}%)`;
-  }
-
-  return {
-    title: {
-      text: title,
-      diff: {
-        text: tooltipDiffText,
-        color: tooltipDiffColor,
-      },
-    },
-    tooltipData: [newLeft, newRight],
-  };
-}
-
-function percentDiff(leftPercent: number, rightPercent: number): number {
-  // difference between 2 percents
-  // https://en.wikipedia.org/wiki/Relative_change_and_difference
-  return ((rightPercent - leftPercent) / leftPercent) * 100;
+  return shouldShowFooter ? (
+    <div className={styles.clickInfo}>{clickInfo}</div>
+  ) : null;
 }
